% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jptransnet-package.R
\docType{package}
\name{jptransnet-package}
\alias{jptransnet}
\alias{jptransnet-package}
\title{jptransnet: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
